finished code, began to add docstrings
