+++ braindecode/augmentation/transforms/global_variables.py
