+++ braindecode/augmentation/transforms/masking_along_axis.py
@@ -4,9 +4,7 @@ from .global_variables import fft_args, data_size
 def mask_along_axis(X, params):
     r"""
     Apply a mask along ``axis``. Mask will be applied from indices
-    ``[v_0, v_0 + v)``, where
-    ``v`` is sampled from ``uniform(0, mask_param)``,
-    and ``v_0`` from ``uniform(0, max_v - v)``.
+    ``[v_start, v_end)``, .
     All examples will have the same mask interval.
 
     Args:
@@ -34,7 +32,19 @@ def mask_along_axis(X, params):
 
 
 def mask_along_axis_random(X, params):
+    """Given a magnitude and an axis, produces a masking interval ``[v_start, v_end)``
+    where
+    ``v_end - v_start`` is sampled from ``uniform(0, magnitude * v_max)`` and rounded,
+    and ``v_start`` from ``uniform(0, v_max - v)`` and rounded.
 
+    Args:
+        X (Tensor): data
+        params (dict): a dict containing the necessary parameters for applying the transform : the magnitude, the axis and the masking value.
+
+    Returns:
+        Tensor: Masked data
+    """
+    
     specgram = X
     value = torch.rand(1) \
         * params['magnitude'] * specgram.size(params["axis"])
@@ -43,14 +53,22 @@ def mask_along_axis_random(X, params):
 
     params["mask_start"] = (min_value.long()).squeeze()
     params["mask_end"] = (min_value.long() + value.long()).squeeze()
-
     X = mask_along_axis(X, params)
     return X
 
 
 def mask_along_time(datum, magnitude):
+    """Given a magnitude and data, will mask a random band of data along the time axis
+
+    Args:
+        datum (Datum): A wrapper containing the data to transform, plus metadata informations useful for certain transforms
+        magnitude (float): a ``[0, 1] float, harmonized between transforms, characterizing how much the transform will alter the data 
+
+    Returns:
+        Datum: A wrapper containing the transformed data.
+    """
     X = datum.X
-    params_time = {"magnitude": magnitude, "axis": 2}
+    params_time = {"magnitude": magnitude, "axis": 2, "mask_value": 0}
     X = signal_to_time_frequency(X)
     X = mask_along_axis_random(X, params_time)
     datum.X = time_frequency_to_signal(X)
@@ -58,8 +76,17 @@ def mask_along_time(datum, magnitude):
 
 
 def mask_along_frequency(datum, magnitude):
+    """Given a magnitude and data, will mask a random band of data along the frequency axis
+
+    Args:
+        datum (Datum): A wrapper containing the data to transform, plus metadata informations useful for certain transforms
+        magnitude (float): a ``[0, 1] float, harmonized between transforms, characterizing how much the transform will alter the data 
+
+    Returns:
+        Datum: A wrapper containing the transformed data.
+    """
     X = datum.X
-    params_frequency = {"magnitude": magnitude, "axis": 1}
+    params_frequency = {"magnitude": magnitude, "axis": 1, "mask_value": 0}
     X = signal_to_time_frequency(X)
     X = mask_along_axis_random(X, params_frequency)
     datum.X = time_frequency_to_signal(X)
@@ -67,6 +94,14 @@ def mask_along_frequency(datum, magnitude):
 
 
 def signal_to_time_frequency(X):
+    """Transforms a temporal signal into its time-frequency representation
+
+    Args:
+        X (Tensor): a temporal signal
+
+    Returns:
+        Tensor: a time-frequency representation
+    """
     global fft_args
     X = torch.stft(X, n_fft=fft_args["n_fft"],
                    hop_length=fft_args["hop_length"],
@@ -75,6 +110,13 @@ def signal_to_time_frequency(X):
     return X
     
 def time_frequency_to_signal(X):
+    """Transforms a time-frequency representation back into a signal
+    Args:
+        X (Tensor): a time-frequency representation
+
+    Returns:
+        Tensor: a signal 
+    """
     global fft_args, data_size
     X = torch.istft(X, n_fft=fft_args["n_fft"],
                     hop_length=fft_args["hop_length"],
